[podspec] Make all subspecs build as frameworks and static libs.
The cxxreact and jschelpers modules don’t need to be exposed publicly in
any way, so the issue was really only with building the internal
sources, for which adding internal header search paths was enough.

This uses `PODS_TARGET_SRCROOT` to ensure that these paths always
resolve correctly regardless of integration type, e.g. with `:path`, but
this has only recently be added to CocoaPods 1.2.0.beta.1